chore: dedup shared logic with legacy path
